+++ packages/implementations/react/src/atomic/Badge/src/badge.types.ts
@@ -1,14 +1,13 @@
 import { ReactNode, HTMLAttributes } from "react";
 import { BadgeVariants } from "@rarui/styles";
 
-export interface Badge {
+export interface BadgeTyping {
   /**
-   * The content of the button.
+   * The content displayed within the badge. This can be text, icons, or any other ReactNode.
    * @TJS-type React.ReactNode
    */
   children: ReactNode;
 }
 
-export type BadgeProperties = BadgeVariants;
-
+export type BadgeProperties = BadgeTyping & BadgeVariants;
 export type BadgeProps = BadgeProperties & HTMLAttributes<HTMLDivElement>;