+++ packages/implementations/react/src/atomic/Input/src/components/Search/search.types.tsx
@@ -1,9 +1,9 @@
 import { InputHTMLAttributes } from "react";
 import { InputProperties } from "../../input.types";
 
-export interface Search {}
+export interface SearchTyping {}
 
-export type SearchProperties = Search &
+export type SearchProperties = SearchTyping &
   Pick<InputProperties, "divider" | "appearance" | "size">;
 
 export type _SearchProps = SearchProperties &