+++ packages/styles/src/packages/atomic/banner/rarui-banner.types.ts
@@ -1,4 +1,4 @@
 import { RecipeVariants } from "@vanilla-extract/recipes";
-import { styles } from "./raui-banner.css";
+import { bannerStyles } from "./raui-banner.css";
 
-export type BannerVariants = RecipeVariants<typeof styles.banner>;
+export type BannerVariants = RecipeVariants<typeof bannerStyles.banner>;