+++ packages/styles/src/packages/atomic/box/index.ts
@@ -1,7 +1,7 @@
-import { sprinkle } from "./rarui-box.css";
+import { boxSprinkle } from "./rarui-box.css";
 
 export const box = {
-  sprinkle,
+  sprinkle: boxSprinkle,
 };
 
 export type { BoxSprinkle } from "./rarui-box.types";