+++ packages/styles/src/packages/atomic/box/rarui-box.css.ts
@@ -153,4 +153,4 @@ const defineProperties = defineRainbowProperties({
   },
 });
 
-export const sprinkle = createRainbowSprinkles(defineProperties);
+export const boxSprinkle = createRainbowSprinkles(defineProperties);