+++ packages/implementations/react/src/atomic/Button/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 The Button component allows us to initiate actions, make state or page changes.
 
+## 2024-05-07 `1.1.1`
+
+#### 🐛 Bug fixes
+
+- Fixed component typing. ([#32](https://git.rarolabs.com.br/frontend/rarui/-/merge_requests/32) by [@junior](https://git.rarolabs.com.br/junior))
+
 ## 2024-05-07 `1.1.0`
 
 #### 🎉 New features