+++ packages/implementations/react/src/atomic/Divider/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 A Divider is a thin line used to separate or group content into lists and layouts.
 
+## 2024-05-07 `1.1.1`
+
+#### 🐛 Bug fixes
+
+- Fixed component typing. ([#32](https://git.rarolabs.com.br/frontend/rarui/-/merge_requests/32) by [@junior](https://git.rarolabs.com.br/junior))
+
 ## 2024-05-07 `1.1.0`
 
 #### 🎉 New features