+++ packages/implementations/react/src/atomic/Title/CHANGELOG.md
@@ -7,6 +7,7 @@ Tilte is a basic component that allows you to give titles and more hierarchy to 
 #### 🐛 Bug fixes
 
 - Fixed component typing property `color`. ([#32](https://git.rarolabs.com.br/frontend/rarui/-/merge_requests/32) by [@junior](https://git.rarolabs.com.br/junior))
+- Fixed component typing. ([#32](https://git.rarolabs.com.br/frontend/rarui/-/merge_requests/32) by [@junior](https://git.rarolabs.com.br/junior))
 
 ## 2024-04-25 `1.0.1`